requirements.txt
google-generativeai>=0.3.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.8
pinecone-client>=2.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
rich>=13.0.0
sqlalchemy>=2.0.0
streamlit>=1.28.0
